clean up cli command structure isolate each cmd in terms of their feature
