+++ crates/fluvio-cli/src/client/derivedstream/delete.rs
