+++ crates/fluvio-cli/src/client/partition/list.rs
