+++ crates/fluvio-cli/src/client/smartmodule/delete.rs
