+++ crates/fluvio-cli/src/client/smartmodule/mod.rs
