+++ crates/fluvio-cli/src/client/tableformat/create.rs
@@ -10,9 +10,12 @@ use tracing::debug;
 
 use fluvio::Fluvio;
 use fluvio::metadata::tableformat::TableFormatSpec;
-use crate::tableformat::TableFormatConfig;
 
 use crate::CliError;
+
+use super::TableFormatConfig;
+
+
 
 // -----------------------------------
 // CLI Options
