+++ crates/fluvio-cli/src/client/tableformat/delete.rs
